backend/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.6.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-cors-headers==3.8.0
django-extensions==3.1.3
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.0.1
drf-yasg==1.20.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
JPype1==1.1.2
konlpy==0.5.2
lxml==4.6.3
MarkupSafe==2.0.1
mysqlclient==2.0.3
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.3.3
protobuf==3.17.3
py4j==0.10.9
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
pyspark==3.1.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scipy==1.7.1
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tweepy==3.7.0
uritemplate==3.0.1
urllib3==1.26.6
xmltodict==0.12.0
wheel